requirements.txt
＃运行环境里所需要的第三方python包

#亚马逊S3服务器的api
boto

#移动端探测
django_mobile

#python与MySQL连接所需要的包
MySQLdb

#python的图片处理包
PIL

#数据库迁移包
south

#json包
simplejson

